requirements.txt
python-telegram-bot[job-queue]==21.6
openai>=1.51.0
python-dotenv>=1.0.0
httpx>=0.27.0
